Rewrite rendering using a command buffer abstraction.
(All managed through the new CommandBuffer class.)

Things that want to draw now populate a provided CommandBuffer using
a fixed set of available rendering commands.  CommandBuffers can be reused
across multiple frames and CommandBuffers can include calls to other
CommandBuffers.

Renderers now include a method to render the contents of a given
CommandBuffer; this in turn helps us isolate all of the OpenGL-isms in the
opengl2 renderer, which in turn makes it (potentially) easier to write
additional renderers--e.g., for Metal or DX.