RadarScopePane: improve label positioning
VOR/NDB labels are now to the left of the point.
Airport labels are centered below.
(And fixes are, as before, to the right.)

This fixes labels being drawn on top of each other when there is,
for example, a VOR coincident with an airport and both labels are
being drawn.