Add CurrentTime() method to ControlServer
Most time-related things are now based on time as specified by
the server, which allows replaying traces with the same (apparent)
time and better handling of sped-up trace replays. (e.g., some bugs
with incorrect-length vector lines are taken care of with this.)

Some time related things are still on the actual time--e.g. timers.