CLI: allow multiple names, generalize track/drop commands
Commands can now have multiple names, so we can have both
longer and shorter variants of them (e.g. both "track" and "tr" to
track.)

Separate RejectHandoff and AcceptHandoff commands are gone;
now the drop track commands drops track if we're tracking and
refuses a handoff if it's being offered. Similarly, track either
initiates track or accepts a handoff, depending on the circumstances.