Rework aircraft selection flow
RadarScope now posts a SelectedAircraftEvent upon a selection

This is now consumed by PositionConfig monitoring the event stream to set its selectedAircraft member variable.

PositionConfig's NotifyAircraftSelected method is now gone, as is the
optional Pane ConsumeAircraftSelection method (which was only used
by the CLI pane)

This does mean that we can no longer click on an aircraft and have its
callsign inserted in the current CLI input, but that functionality will shortly
be replaced with the f-key / status bar approach.