Rework CLICommand interface
Ditch the CommandArgsFormat stuff; now commands can indicate
that they want an aircraft, a controller, and some range of number
of arguments. The aircraft now must be the selected aircraft and
the controller, if required, is taken from the first argument.

This simplifies both CLICommand implementations as well as the
CLIPane that gets things organized and runs commands. We lose
the chance to specify an unselected aircraft's callsign in a command
entered by the keyboard, though this seems like a minor loss for
a cleaner model.