Rework imgui draw command handling
The Renderer interface no longer has a specialized RenderImgui
method; rather, a new GenerateImguiCommandBuffer converts
imgui draw command lists into calls into vice CommandBuffers.

This change required generalizing how CommandBuffer represents
vertex arrays so that it could support the interleaved arrays that
imgui generates (as well as the packed 8-bit RGBAs that it uses
for colors.)

The main advantage of this is that it's that much easier to add a new
Renderer backend; now all it needs to do now is to be able to turn
CommandBuffers into pixels.