Big rework of status bar / f-key command handling
F-key commands are handled in the CLIPane now; this lets us
simplify the whole pane-focus model so that it doesn't have
special cases for the status bar.

The status bar at the top is now used only for general system
error messages and isn't visible otherwise.

F-key bindings are now stored in the PositionConfig rather than
GlobalConfig, which gives a little more flexibility in setting them.
Thus, any f-key bindings previously set in config files will be lost
with this change.