Improve handling of canceled handoffs
Distinguish between handoffs we offered and another controller rejected
and handoffs another controller offered and then rescinded in handling
the #PC::CCP:HC FSD message.

Added a CanceledHandoffEvent for the event stream (and pay attention
to it as appropriate.)

In the CLI's TrackAircraftCommand, issue a cancel for an aircraft that
we're trying to hand off.