Add multi- drop and track commands for f-keys
These do different things depending on context, which allows doing
more with a single f-key binding.

Track will track an untracked aircraft, accept an offered handoff from
another controller, or rescind a handoff that we've offered to another
controller. Drop either drops an aircraft that we're tracking or rescinds
the offer of a handoff to another controller.