Fix bug in updating waypoints after approach clearance
Previously, we would add all waypoints that were closer to the airport than the
aircraft was. However, that would sometimes lead to a waypoint behind the
aircraft being added, which in turn led to the aircraft basically floating in space
while it tried to go back to the previous waypoint. Now we also check that the
waypoint is actually in front of the aircraft as well.