Clean up ScenarioGroup state in SimConnectionConfiguration
Keep a *ScenarioGroup in SimConnectionConfiguration and use it
for all of the UI stuff, rather than setting the global scenarioGroup
variable, which may end up being the wrong thing anyway if the user
hits cancel in the connection dialog. Now the global is only set (and
the STARSPane notified) once it's a done deal.