Rework handling of arrivals with different routes based on runway
Now aircraft start out with the common waypoints at the start of the
approach when they spawn. Once "expect" is given for an arrival, we know
the runway and then the remaining arrival waypoints for that runway are
added to the route.

As part of this, the JSON specification has changed so that in an arrival,
"waypoints" must always be specified. Then, "runway_waypoints" has only the
runway-specific suffixes (with the last "waypoints" waypoint required to be
the same as each initial "runway_waypoints" waypoint..)