Maintain separate tracking controllers and controlling controllers
This way we can hand off an aircraft, have the other controller accept the
track, but still issue control instructions to it. After the track is clicked to
acknowledge the handoff (making it go from white to green), then the other
controller has control as well. Then only at that point does it climb the aircraft
to its cruising altitude...