Fix final approach speed on overflights
Previously FinalApproachSpeed slowed the aircraft purely based on its
distance to the airport; this led to very slow overflights (e.g. when an aircraft
was passing near the airport on its way to a procedure turn.) Now it computes
the flying distance to the airport and also makes sure the aircraft is generally
pointing at it before it starts slowing it down.