Fix bugs in Aircraft ShouldTurn* methods that affected original Aircraft state
Because the AssignHeading method was called with ac2, it ended up canceling
the approach clearance, which in turn affected the original Aircraft since the
map was shared between them. Now we just directly set Nav.L for the turn and
duplicate the maps; this should have the further advantage of correctly
modeling any future nav commands that would kick in during the turn.