Restructure handling of RunAircraftCommands
Now, the string goes all the way to the server, rather than being turned into
a string of RPC calls. Not only is this faster, but it also fixes a bug where a
failing command wouldn't prevent the subsequent ones from running.

As a bonus, this now fits nicely into the structure of other commands, so the
PendingAircraftCommands struct is no longer needed.