Improve elapsed time handling in Sim Update()
We now track the elapsed time that hasn't been accounted for by Sim updates:
the Sim runs at 1 Hz tick rate, so if, for example, 1.1s go by between updates,
we should run the sim for one tick and remember that 0.1s were unaccounted
for. Then, we only need 0.9s more to go by before it's time to run the Sim's
next tick.

In general usage, this doesn't matter much, but it starts to become a factor at
high SimRates, in which case we weren't actually running at the requested rate.
It also lets us now sleep for 100ms between updates without losing time with
a SimRate of 1.