Remove lastDeparture from ScenarioGroupDepartureRunway
That was the wrong place for it (that being a place for read-only data that
defines the sim.) Now the Sim manages this information directly and passes
it in to World CreateDeparture.

Along the way, fixed a bug in CreateDeparture where when the exact same
departure was reused, we'd spawn an aircraft to the exact same destination.
Now, we sample from departures with exit fixes that match the previous
departure
in that reuse case...