Remove RadarTrack array from Aircraft; maintain tracks in STARSAircraftState
This is a better place for them to live, since they are not so much inherent to
an aircraft, but to its being seen by a radar. Moving them here will make it
easier to update them according to the radar sweep in the future.

This has the added benefit of reducing network bandwidth for aircraft updates. (Related, bumped the value of ViceRPCVersion.)