Improve STARSPane flight plan display
Per the STARS manual, proposed departures, active departures, and arrivals all
have different formats, which are now (mostly) handled. We do not include the
third line for any of them, though, which is all about mode S transponder
stuff that we don't support.

As part of this, autotrack now only initiates track when an aircraft is first
visible to radar, which seems better than initiating track immediately when
it spawns...