Remove ArrivalRunwayWaypoints from Aircraft
Instead store the arrival group and an index and then grab the runway waypoints
when needed from World. No change to functionality, but makes the Aircraft struct
less unwieldy when printed out for debugging.

(Also requires an RPC and serialization version bump, though the one in af58daddbb9
will do.