Add {Departure,Arrival}AirportElevation to FlightPlan
This information is useful here and there, so it's nice to not need to explicitly look it up.
Added a NewFlightPlan function that takes care of initializing this information
(and airport locations) for spawned aircraft.

(This also is RPC/serialization incompatible, so it's a good time to land this now.)