Multiple improvements to pilot radio comms
There is a new RadioTransmission type that wraps up the message, target controller,
and the type of transmission, which may either be a readback or a pilot-initiated
communication (e.g. checking in after a hand off, or informing the controller of
a go around.) All of the Aircraft action methods now return a []RadioTransmission
in place of the two strings used for readbacks.

Other polish:
- Capitalization is fixed
- There's now a bit of random variation in the pilot responses
- Pilots report "unable" in more error cases, like direct to a fix that isn't in
  their route
- Fixed FormatAltitude() for altitudes <1000'
- Improved NavSTATE ContactMessage and added AltitudeMessage (for departures)