Fix bug in GoAround
Using the distance to the airport is a bad idea for short distances, since the
distance at which the go around happens is <1 mile. Thus, sometimes aircraft that
were supposed to go around never would, if the arrival runway threshold was
far enough away from the airport center.

Now Aircraft FinalApproachDistance is used for the distance-to-airport determination.