Fix bug that would lead to speed restrictions on final being ignored
Basically, we'd have a residual route in Aircraft Waypoints from the STAR
(e.g., a single waypoint), an aircraft on a heading to intercept but without
the rest of the approach waypoints in its Waypoints, would decide it was close
to the airport, and ApproachSpeedat5DME would kick in.

Now, Aircraft FinalApproachDistance() just gives up and returns an error if
the LNav mode isn't FlyRoute, and ApproachSpeedat5DME doesn't do its thing
until it gets an error-free distance to the airport that's also under 5 nm.