Initial rework of the navigation system
The whole idea of pluggable interfaces is gone; we ended up with lots of
type checks and special case handling of navigation state that is more
cleanly handled with an explicit state representation, along the lines of
what we had before 36e9f40, but hopefully improved this time around.

Now aircraft navigation, including its current location, heading, etc., is
centralized in the Nav class, which has also taken over the Aircraft
methods related to updating the aircraft flight state.

HILPTs are not yet wired up. There are likely many other bugs in this version.