Rework and generalize handing of departures
Control positions in multi_controllers can now list the airports (and optionally
runways) for which they are the departure controller. As before, if the controller
isn't signed on, vice works through the backups until it finds one.

Also moved handling of passed waypoints from Aircraft Update() to Sim
updateState(), which is a better place for it and lets us fix not knowing
which controller to use for handoffs in the Aircraft code and needing to
patch it up in Sim.