Improve handling of /ho in routes
In particular, this makes the not-often-used case of departures that initially
go to a virtual controller and are then handed off work better.

Aircraft ArrivalHandoffController is now renamed to WaypointHandoffController.
It is initialized as before for arrivals but now for departures, the controller
comes from the ExitRoute. As part of wiring this up, some of the details of
virtual vs human controllers and Aircraft iniitalization moved from world.go to
aircraft.go.

Updated the JFK scenario accordingly (for LGA->JFK departures).

This necessitated a bump of the RPC and config serialization version numbers.