STARSPane: rework handling of tower lists
A number of scenarios had the same index for tower_list, which led to multiple
lists being drawn on top of each other. Now tower_list is treated as a priority
index and the three tower lists are handed out based on tower_list and then
alphabetically by airport names in case of a tie. In this way we ensure that
the tower lists are all allocated to airports in the scenario (given enough
airports.)