Move various STARS-related config items from ScenarioGroup to STARSFacilityAdaptation
Now that we have that (via "stars_config"), it makes more sense to have things
like the radar scope default center, the inhibit CA volumes, etc., to live under it.
Also updated all of the scenarios accordingly.

The data layout is now quite incompatible with the last release, so bumped the
config file version and RPC version here as well.