CommandBuffer: clamp width and height in SetDrawBounds to 0
Dragging the bottom of the STARS window up into the DCB led to a degenerate
Extent2D being passed into CommandBuffer SetDrawBounds; in turn, that led
to bogus values going into the command buffer and, presumably, invalid
values being passed to glViewport and glScissor.

This matches the description of #121 and likely fixes it though it's hard
to be sure, not knowing exactly what triggered it.