STARSPane: fix bugs in handling of track history
Now we do a better job of distinguishing between the last track from the radar
(which is now stored in STARSAircraftState track) and the track history,
which is stored in STARSAircraftState historyTracks. We follow STARS better
in that there is now a history update rate (set via the H_RATE DCB button) and
the history tracks are updated the first time we get a new radar track and the
history rate number of seconds have passed.

Making this distinction for example lets us use the most recent radar track
altitude in datablocks without regard to the track history (and its
update rate.)

Also fixed history drawing to not draw history tracks at all if the associated brightness
is zero.

Fixes issue #75.