Fix regression in handling of track history
The updated handling of track history in 8460812c10838 inadvertently
introduced a regression, where (for example) the PTLs for aircraft in a turn would
only be updated at the history rate (~5s) and not every time the radar supplied us
with a new track (~1s).

With this fix now we always store the most recent and the second most recent
returns from the radar regardless of the track history rate used to display track
tails. These are then used to calculate the aircraft heading and such, which
fixes the regression...