Switch to (apparently) real-world STARS fonts
This includes an update to the Font structure to store the texture id of
the atlas (no more assumption of a single atlas) as well as corresponding
changes to TextDrawBuilder to use this.

The STARS fonts are converted from PCF bitmap files using util/pcf2go.cpp,
which is a hacked up version of https://github.com/ganaware/pcf2bdf that
instead generates a golang source file that stores the bitmaps and
associated font information in structs.

One nit is that we don't have the font for the smallest size, so that is
generated by downscaling the bitmap of the smallest one we do have.