WIP support for directly loading DAT video maps
TODO: error handling, comments, more robust culling of the border circle and text.

See videomaps/a90.zip for an example. Each set of videomaps (e.g. for a TRACON) goes in a ZIP file where each map is its own *.dat file. The ZIP file should have a maps.json file in it that gives the name for each map in the file and its map number.

Note that the a90.json scenario has been butchered to test this with all scenarios using the single map.