changed how squawks are assigned
Squawks are now assigned with Aircraft NewFlightPlan as assigned squawks are now held in the FlightPlan section of the Aircraft struct. In the future, STARS will be more reliant on information from the flight plan, and will eventually associate the flight plan to an aircraft instead of having NewFlightPlan do it.