Fix sporadic "fix _SLOAP not found in navaid database" error
At startup / validation time, sometimes but not always we would get the
above error.

After much debugging, it became apparent that the issue was that the STAR
a) came from the FAA CIFP database, b) had multiple transitions with a
given fix, c) was used multiple times in a scenario, and where d) said fix
was used as a "spawn" point for one of the uses of the STAR.

The issue, then, was that the code that inserted the spawn point (with the
placeholder fix name with a "_" prepended) to the route didn't make a copy
of the underlying slice of Waypoints. In turn, when that slice was used in
other cases, we'd have the spawn point in the route which then woudln't be
found in the fix database.

Good times.