Rework ERAM/STARS computer structure
ERAMComputers no longer hold pointers to the other ERAMComputer inboxes,
and similarly for the STARSComputers. Each ERAMComputer instead holds a
pointer to the top-level ERAMComputers object, from which it can find its
way to the other computers, when they are needed.

The old pointer-heavy way of making each computer aware of the others that
it might need to talk to led to problems with serialization (i.e., a "gob:
encodeReflectValue: nil element: error, which prevented loading the old
session after quitting and restarting vice.