sim.State: rework video map handling
Previously the client would get video maps from the STARSFacilityAdaptation
in sim.State, which ends up holding all of the video maps for all of the
controllers in a scenario (i.e., is fairly wasteful.) Now we store only the
ones the controller needs for their actual position in the State object.

Not only does this save some network bandwidth when connecting to a Sim,
but it also reduces server memory use since encoding/gob seems to like
holding on to its temporary buffers, so a one-time transmit of something
big balloons up memory use.