STARSPane: rework handling of video maps
(This was all necessary to fix a bug where nothing was drawn when one of
the PHL ATCT maps was selected, in the end due to inconsistent handling of
video map renumbering when multiple shared the same id.)

The STARSPane now keeps its own copy of the slice of video maps returned by
sim.State GetVideoMaps() so that when it renumbers maps with repeated ids,
that renumbering isn't lost to other parts of the STARSPane implementation.

Also, moved video map drawing to a separate function.