STARS: fix bugs in handling of fonts on high-DPI windows systems
We were drawing too-small quads but then taking the correct size steps in x and y,
so spacing was terrible and the fonts were too small.

As part of fixing that, simplified font handling by no longer doubling up the font bitmap
for Windows high-DPI systems; there was no need for that (given nearest sampling) and
drawing properly-sized quads is sufficient.