util: add CatchAndReportCrash and crash upload server support
CatchAndReportCrash is designed to be called with a defer statement at the
start of most goroutines and centralizes handling of panics. In addition to
checking if running under the debugger, it logs the stack trace and some
basic info to disk and also attempts to upload it to the vice server, which
now listens on a port for crash reports and saves them to disk.

Also has the addition of defer'ed calls to this from a variety of places
that hadn't caught and logged panics before, including (unfortunately) most
of the sim.Dispatcher methods (since the RPC runtime spins up goroutines
for those as messages come in.)