sim: add wait before pilots contact controllers / depart on course
We try to model the time taken to change the frequency on the radio and get
around to calling in; in the interim between when the first controller
issues the "contact xyz" instruction and the aircraft calls in to the new
controller, no one is able to issue control instructions.

Related, when a departure is handed off to a virtual controller, there is a
wait of a little while before they proceed to their first fix and climb to
cruising altitude, again to better model real-world behavior.