math: take handle magnetic variation in Offset2LL
Of the two callers, stars parseRALocation was doing the correction manually,
so it should return the same results. The other is scenario files that use FIX@HDG/DIST;
these will give slightly different results before. Currently only AAC seems to use
this, and a quick eyeball of the updated routes looks reasonable.