sim: refactor into 'server' and 'sim' packages.
Hopefully no functional changes, though a bunch of direct accesses to the
SimManager controllerTokenToSim map now go through the ControllerTokenToSim
method, which grabs the mutex first. This may be the cause of some prior
mysterious crashes in e.g. Dispatcher TogglePause() (and hopefully will not
introduce deadlocks due to the mutex already held by callers, though that
shouldn't be an issue for the dispatcher.