Rework controller statistics measurement
We now track them in ControlClient, purely for the current controller (and
restarting at zero upon a restart.) Counters are only incremented when the
controller initiates track or accepts a handoff; it's no longer effectively
the total number of aircraft spawned ever.

This will hopefully give a good foundation for expanding the stats that are
tracked.