Updated approach for low-latency state updates
Rather than locally updating the state as a band-aid until the next proper state
update from the server, now many of the RPC calls generate and return a StateUpdate
object directly; this is then applied immediately when it's received. While this has
higher worst-case latency than before, it simplifies the code overall, eliminating
the need for duplicate logic for the quick local and then canonical remote state
updates.

Other changes:
- WorldUpdate ->StateUpdate ("world" is stale / n/a in the current system.)
- RPCs take a single callback that is passed the error return value which is
  nil on success.
- Moved RPCClient/PendingCall from util to server
- Removed SetGlobalLeaderLine RPC (unused; that's now done via flight plan
  modification.)
- GetStateUpdate no longer does a deep copy of the StateUpdate; I don't
  think this is actually necessary even with a local server since it's being sent
  over RPC in any case.