av: distinguish between transponder-reported and true alt. in RadarTrack
TrueAltitude is always the actual altitude; it should be used in cases like checking
whether a track is visible to a RADAR. TransponderAltitude is the altitude the
transponder is reporting which is 0 if the transponder is not squawking
altitude. (In the future we might use this for broken/inaccurate transponders.)

All code that references track altitudes has been updated to (hopefully)
use the appropriate one.

This fixes a bug where telling a departure that was still close to the airport to
squawk standby would cause them to disappear from the scope.