nav: rework handling of turns
There's now a simple model of aircraft banking and turn rate now also
depends on the aircraft speed. This leads to more of an ease-in/ease-out
for turns and also means that if they're turning in one direction and then
vectored in the other direction, there's a transition phase where they keep
turning the original direction until they get level and then start the turn
in the new direction.