nav: add missing case for already being at a waypoint speed restriction
Not handling this (and not returning from TargetSpeed() at that point) caused us
to fall through and try to fly the speed from targetAltitudeIAS() -> this caused
deceleration and so slowed the descent rate. Then the next tick we'd see that
we were not meeting the speed restriction and would accelerate,
again slowing
the descent rate. In turn, aircraft were massively missing altitude restrictions.