sim: clean up time handling for prespawn
Previously, SimTime would be set to time.Now() initially, other things would be
initialized based on this (including WX), and then Prespawn() would roll back
the time according to the prespawn period. This was a little ugly and, more
importantly, broke the wind gust calculations due to times unexpectedly
far in the past. Now we set SimTime to include that rollback from the start;
this also allows cleaning up some of the prespawn flow.