Add support for STARS consolidation commands
This required a significant rework of track ownership and user session
management, but cleaned up a lot of things in the end. The sim/ layer is no
longer aware of user sign ins / sign offs; that all happens at the server/
layer. The Sim just knows about workstations and what positions each one is
responsible for, independent of whether someone is sitting in front of one.

This also allowed cleaning up the instructor/RPO role implementations as
well as letting users sign in to an occupied workstation to relieve another
controller.

Fixes #227 and #381.