SNA Update (Ketan K)
For SNA, also just some minor tweaks and QOL improvements for the scenarios. I've removed the clearance for the RNAV Z at JWARD being given by Area 3, as IRL, most of the time, aircraft are manually vectored for the ILS or RNAV by either Maverick or Tustin for the sake of simplicity.