sim: clean up State updates with new UpdatedState type
This more explicitly separates what is static in State and what is sent to the user
per-tick. In turn, the update code is simpler, just copying the full struct.
As part of this, moved some housekeeping code out of GetStateUpdate
and into the per-second sim update loop.

Also removed ACFlightPlans from StateUpdate; it was only used by FlightStripPane
and the info it required was easily enough available elsewhere.